Extract pure helpers + constants from create-poll page
Move pollDataToMultipollRequest, validateRankedChoiceOptions, and the
deadline option arrays (BASE/FRACTIONAL/ABSOLUTE/DEV_DEADLINE_OPTIONS)
out of page.tsx and into createPollHelpers.ts. They're pure / data-only
and were just buffering page.tsx with module-local noise.

app/create-poll/page.tsx: 1870 -> 1755 lines.
